+++ private/AdfPublishOption.class.ps1
@@ -1,7 +1,7 @@
 class AdfPublishOption {
     [hashtable] $Includes = @{}
     [hashtable] $Excludes = @{}
-    [Boolean] $DeleteNotInSource = $true
+    [Boolean] $DeleteNotInSource = $false
     [Boolean] $StopStartTriggers = $true
     [Boolean] $CreateNewInstance = $true
 }