+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Incoming version
+### Changed
+* DeleteNotInSource option is $false by default now
+* 
+
 ## [0.9.0] - 2020-06-03
 ### Added
 * Additional method of deployment by New-AzResource cmdlet (default)